+++ commerce/warehouse/src/main/java/ru/yandex/practicum/controller/WarehouseController.java
@@ -41,7 +41,7 @@ public class WarehouseController implements WarehouseClient {
     }
 
     @Override
-    public void productToWarehouse(Map<UUID, Integer> products) {
+    public void productToWarehouse(Map<UUID, Long> products) {
         service.productToWarehouse(products);
     }
